backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.19
psycopg[binary]>=3.2.0
PyJWT>=2.8.0
pydantic-settings>=2.3.0
alibabacloud-sts20150401>=1.1.6
alibabacloud-tea-openapi>=0.3.12
oss2>=2.18.6
